feat: adapt ZERO_HASHES for poseidon2
